requirements.txt
fastapi
pydantic
uvicorn
python-multipart
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.4,<1.8.0
sqlalchemy
langchain
langchain-community
langchain-core
langchain-openai
openai
faiss-cpu
python-dotenv
gunicorn
tiktoken
pypdf
pdfminer.six
chromadb